Nav | Eliminating errors, cleanup
