+++ resources/assets/js/store/modules/course.js
@@ -12,6 +12,10 @@ function getCourseApiUrl(courseId) {
 	)
 }
 
+const STATUS_NONE = 'none'
+const STATUS_IN_PROGRESS = 'in-progress'
+const STATUS_AVAILABLE = 'available'
+
 // Namespace
 const namespaced = true
 
@@ -94,6 +98,32 @@ const getters = {
 
 		return undefined
 	},
+	nextLesson: (state, getters, rootState, rootGetters) => {
+		let lesson = { status: STATUS_NONE },
+			inProgressId = rootGetters['progress/getFirstLessonIdInProgress'](state.id)
+
+		if (inProgressId > 0) {
+			lesson = getters.getLesson(inProgressId)
+			lesson.status = STATUS_IN_PROGRESS
+		} else {
+			for (var lessonId in this.getLessons) {
+				let isAvailable = rootGetters['progress/isLessonAvailable'](lessonId)
+				if (isAvailable &&
+					!rootGetters['progress/wasLessonStarted'](state.id, lessonId)
+				) {
+					lesson = this.getLesson(lessonId)
+					lesson.status = STATUS_AVAILABLE
+					break
+				} else if (!isAvailable) {
+					lesson = this.getLesson(lessonId)
+					lesson.status = STATUS_NONE
+					break
+				}
+			}
+		}
+
+		return lesson
+	}
 }
 
 // Mutations
